Basic server-side user and group search API

+++ iam-login-service/src/main/java/it/infn/mw/iam/api/tokens/service/DefaultRefreshTokenService.java
@@ -72,7 +72,7 @@ public class DefaultRefreshTokenService extends AbstractTokenService<RefreshToke
     return tokenRepository.findAllValidRefreshTokens(new Date(), op);
   }
 
-  private int countAllValidTokens() {
+  private long countAllValidTokens() {
 
     return tokenRepository.countValidRefreshTokens(new Date());
   }
@@ -83,7 +83,7 @@ public class DefaultRefreshTokenService extends AbstractTokenService<RefreshToke
     return tokenRepository.findValidRefreshTokensForUser(userId, new Date(), op);
   }
 
-  private int countAllValidTokensForUser(String userId) {
+  private long countAllValidTokensForUser(String userId) {
 
     return tokenRepository.countValidRefreshTokensForUser(userId, new Date());
   }
@@ -94,7 +94,7 @@ public class DefaultRefreshTokenService extends AbstractTokenService<RefreshToke
     return tokenRepository.findValidRefreshTokensForClient(clientId, new Date(), op);
   }
 
-  private int countAllValidTokensForClient(String clientId) {
+  private long countAllValidTokensForClient(String clientId) {
 
     return tokenRepository.countValidRefreshTokensForClient(clientId, new Date());
   }
@@ -105,12 +105,12 @@ public class DefaultRefreshTokenService extends AbstractTokenService<RefreshToke
     return tokenRepository.findValidRefreshTokensForUserAndClient(userId, clientId, new Date(), op);
   }
 
-  private int countAllValidTokensForUserAndClient(String userId, String clientId) {
+  private long countAllValidTokensForUserAndClient(String userId, String clientId) {
 
     return tokenRepository.countValidRefreshTokensForUserAndClient(userId, clientId, new Date());
   }
 
-  private ListResponseDTO<RefreshToken> buildCountResponse(int countResponse) {
+  private ListResponseDTO<RefreshToken> buildCountResponse(long countResponse) {
 
     return new ListResponseDTO.Builder<RefreshToken>().totalResults(countResponse)
         .resources(Collections.emptyList()).startIndex(1).itemsPerPage(0).build();
@@ -129,7 +129,7 @@ public class DefaultRefreshTokenService extends AbstractTokenService<RefreshToke
 
     if (isCountRequest(pageRequest)) {
 
-      int count = countAllValidTokens();
+      long count = countAllValidTokens();
       return buildCountResponse(count);
     }
 
@@ -144,7 +144,7 @@ public class DefaultRefreshTokenService extends AbstractTokenService<RefreshToke
 
     if (isCountRequest(pageRequest)) {
 
-      int count = countAllValidTokensForUser(userId);
+      long count = countAllValidTokensForUser(userId);
       return buildCountResponse(count);
     }
 
@@ -159,7 +159,7 @@ public class DefaultRefreshTokenService extends AbstractTokenService<RefreshToke
 
     if (isCountRequest(pageRequest)) {
 
-      int count = countAllValidTokensForClient(clientId);
+      long count = countAllValidTokensForClient(clientId);
       return buildCountResponse(count);
     }
 
@@ -174,7 +174,7 @@ public class DefaultRefreshTokenService extends AbstractTokenService<RefreshToke
 
     if (isCountRequest(pageRequest)) {
 
-      int count = countAllValidTokensForUserAndClient(userId, clientId);
+      long count = countAllValidTokensForUserAndClient(userId, clientId);
       return buildCountResponse(count);
     }
